+++ gg02/src/styles.scss
@@ -1,5 +1,6 @@
 $light-grey: #b5c6df;
 $bg-color: #282c34;
+$nav-color: #4f5879;
 $neon-blue: rgb(29, 238, 238);
 
 h1 {
@@ -15,5 +16,5 @@ h1 {
   align-items: center;
   justify-content: center;
   font-size: 30px;
-  color: $neon-blue;
+  color: whitesmoke;
 }
